Move write function out of impl Bank
