+++ runtime/src/bank.rs
@@ -217,8 +217,7 @@ struct VerifyAccountsHashConfig {
 }
 
 mod address_lookup_table;
-mod bank_hash_details;
-use bank_hash_details::BankHashDetails;
+pub mod bank_hash_details;
 mod builtin_programs;
 pub mod epoch_accounts_hash_utils;
 mod metrics;
@@ -8213,41 +8212,6 @@ impl Bank {
             }
         }
         false
-    }
-
-    /// Output the components that comprise bank hash
-    pub fn write_hash_details_file(&self) -> std::result::Result<(), String> {
-        let details = BankHashDetails::try_from(self)?;
-
-        let slot = details.slot;
-        let hash = &details.bank_hash;
-        let file_name = format!("{slot}-{hash}.json");
-        let parent_dir = self
-            .rc
-            .accounts
-            .accounts_db
-            .get_base_working_path()
-            .join("bank_hash_details");
-        let path = parent_dir.join(file_name);
-        // A file with the same name implies the same hash for this slot. Skip
-        // rewriting a duplicate file in this scenario
-        if !path.exists() {
-            info!("writing details of bank {} to {}", slot, path.display());
-
-            // std::fs::write may fail (depending on platform) if the full directory
-            // path does not exist. So, call std::fs_create_dir_all first.
-            // https://doc.rust-lang.org/std/fs/fn.write.html
-            _ = std::fs::create_dir_all(parent_dir);
-            let file = std::fs::File::create(&path).map_err(|err| {
-                format!(
-                    "Unable to create bank hash file at {}: {err}",
-                    path.display()
-                )
-            })?;
-            serde_json::to_writer_pretty(file, &details)
-                .map_err(|err| format!("Unable to write bank hash file contents: {err}"))?;
-        }
-        Ok(())
     }
 }
